move dashboard controller to own package
