+++ core/src/main/java/com/hlag/oversigt/controller/EventSourceInstanceController.java
@@ -1,0 +1,5 @@
+package com.hlag.oversigt.controller;
+
+public class EventSourceInstanceController {
+
+}
