+++ core/src/main/java/com/hlag/oversigt/core/OversigtModule.java
@@ -49,11 +49,11 @@ import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import com.google.inject.name.Named;
 import com.google.inject.name.Names;
+import com.hlag.oversigt.controller.DashboardController;
 import com.hlag.oversigt.core.event.EventSender;
 import com.hlag.oversigt.core.eventsource.EventSourceStatisticsManager;
 import com.hlag.oversigt.core.eventsource.NightlyDashboardReloaderService;
 import com.hlag.oversigt.core.eventsource.NightlyEventSourceRestarterService;
-import com.hlag.oversigt.model.DashboardController;
 import com.hlag.oversigt.model.EventSourceNameGenerator;
 import com.hlag.oversigt.properties.Color;
 import com.hlag.oversigt.properties.SerializablePropertyController;
