+++ core/src/main/java/com/hlag/oversigt/core/OversigtServer.java
@@ -54,6 +54,7 @@ import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Singleton;
 import com.google.inject.name.Named;
+import com.hlag.oversigt.controller.DashboardController;
 import com.hlag.oversigt.core.OversigtConfiguration.HttpListenerConfiguration;
 import com.hlag.oversigt.core.WroManagerFactory.CustomWroConfiguration;
 import com.hlag.oversigt.core.WroManagerFactory.WroGroupContent;
@@ -61,7 +62,6 @@ import com.hlag.oversigt.core.event.EventSender;
 import com.hlag.oversigt.core.event.JsonEvent;
 import com.hlag.oversigt.core.event.OversigtEvent;
 import com.hlag.oversigt.model.Dashboard;
-import com.hlag.oversigt.model.DashboardController;
 import com.hlag.oversigt.properties.SerializableProperty;
 import com.hlag.oversigt.sources.MotivationEventSource;
 import com.hlag.oversigt.util.ClassPathResourceManager;
