+++ core/src/main/java/com/hlag/oversigt/web/resources/DashboardResource.java
@@ -47,8 +47,8 @@ import org.slf4j.LoggerFactory;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import com.hlag.oversigt.controller.DashboardController;
 import com.hlag.oversigt.model.Dashboard;
-import com.hlag.oversigt.model.DashboardController;
 import com.hlag.oversigt.model.Widget;
 import com.hlag.oversigt.security.Authenticator;
 import com.hlag.oversigt.security.Principal;
