+++ core/src/main/java/com/hlag/oversigt/web/resources/DashboardWidgetResource.java
@@ -38,8 +38,8 @@ import javax.ws.rs.core.UriInfo;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import com.hlag.oversigt.controller.DashboardController;
 import com.hlag.oversigt.model.Dashboard;
-import com.hlag.oversigt.model.DashboardController;
 import com.hlag.oversigt.model.EventSourceProperty;
 import com.hlag.oversigt.model.Widget;
 import com.hlag.oversigt.properties.Color;
