+++ core/src/main/java/com/hlag/oversigt/web/resources/EventSourceInstanceResource.java
@@ -44,14 +44,14 @@ import org.slf4j.LoggerFactory;
 import com.google.common.base.Strings;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
+import com.hlag.oversigt.controller.DashboardController;
+import com.hlag.oversigt.controller.EventSourceKey;
+import com.hlag.oversigt.controller.InvalidKeyException;
 import com.hlag.oversigt.core.eventsource.EventSourceStatisticsManager;
 import com.hlag.oversigt.core.eventsource.EventSourceStatisticsManager.RunStatistic;
-import com.hlag.oversigt.model.DashboardController;
 import com.hlag.oversigt.model.EventSourceDescriptor;
 import com.hlag.oversigt.model.EventSourceInstance;
-import com.hlag.oversigt.model.EventSourceKey;
 import com.hlag.oversigt.model.EventSourceProperty;
-import com.hlag.oversigt.model.InvalidKeyException;
 import com.hlag.oversigt.security.Principal;
 import com.hlag.oversigt.security.Role;
 import com.hlag.oversigt.util.JsonUtils;
