+++ core/src/main/java/com/hlag/oversigt/web/resources/SystemResource.java
@@ -44,11 +44,11 @@ import com.google.common.collect.EvictingQueue;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.google.inject.name.Named;
+import com.hlag.oversigt.controller.DashboardController;
 import com.hlag.oversigt.core.Oversigt;
 import com.hlag.oversigt.core.OversigtConfiguration;
 import com.hlag.oversigt.core.event.EventSender;
 import com.hlag.oversigt.core.event.OversigtEvent;
-import com.hlag.oversigt.model.DashboardController;
 import com.hlag.oversigt.model.EventSourceInstance;
 import com.hlag.oversigt.properties.SerializablePropertyController;
 import com.hlag.oversigt.security.Authenticator;
